backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
PyMuPDF>=1.23
reportlab>=4.0
Pillow>=10.0
PyPDF2>=3.0
python-docx>=1.0
openpyxl>=3.1
celery>=5.3
redis>=5.0
python-pptx>=0.6.23
python-dotenv>=1.0
